# server/common/swagger/Api.yaml
swagger: '2.0'
info:
  description: Api specs for ICDating
  version: 1.0.0-b1
  title: IC Dating API
  contact:
    email: charlie@aworkingapp.com
basePath: /api/v1
...
schemes:
  - http
paths:
...
  /account/{accountId}:
    get:
      tags:
      - account
      summary: Get account info by account ID
      operationId: getAccountByUserId
      produces:
        - application/json
      parameters:
      - name: accountId
        in: path
        description: The account that needs to be fetched.
        required: true
        type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/AccountVM'
        400:
          description: Invalid username supplied
        404:
          description: User not found
    put:
      tags:
      - account
      summary: Updated account account
      description: This can only be done by the logged in account.
      operationId: updateAccount
      produces:
      - application/json
      parameters:
      - name: accountId
        in: path
        description: acountId that need to be updated
        required: true
        type: string
      - name: token
        description: The web token returned at login or registration
        in: header
        required: true
        type: string
      - in: body
        name: body
        description: Updated account object
        schema:
          $ref: '#/definitions/AccountUpdate'
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/AccountVM'
        400:
          description: Invalid account supplied
        404:
          description: Account not found
  /account/{accountId}/image:
    put:
      tags:
      - account
      summary: Add a new picture url to the account
      operationId: addPhoto
      produces:
        - application/json
      parameters:
      - name: accountId
        in: path
        description: The account that needs to be fetched.
        required: true
        type: string
      - name: token
        description: The web token returned at login or registration
        in: header
        required: true
        type: string
      - in: body
        name: body
        description: The picture url to be added
        schema:
          type: object
          required:
            - pictureUrl
          properties:
            pictureUrl:
              type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              pictureUrl:
                type: string
        401:
          description: Invalid token
        404:
          description: User not found
    delete:
      tags:
      - account
      summary: delete picture url from the account
      operationId: addPhoto
      produces:
        - application/json
      parameters:
      - name: accountId
        in: path
        description: The account that needs to be fetched.
        required: true
        type: string
      - name: token
        description: The web token returned at login or registration
        in: header
        required: true
        type: string
      - in: body
        name: body
        description: The picture url to be removed
        schema:
          type: object
          required:
            - pictureUrl
          properties:
            pictureUrl:
              type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              pictureUrl:
                type: string
        401:
          description: Invalid token
        404:
          description: User not found
...
  /account/hide/initiator/{initiatorAccountId}/receiver/{receiverAccountId}:
    post:
      tags:
        - discovery
      summary: Hide the receiver from the initiator.
      operationId: hideUser
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - name: initiatorAccountId
        in: path
        description: The user who perform the action of hiding someone in their discovery list
        required: true
        type: string
      - name: receiverAccountId
        in: path
        description: The user who received the hiding action and being hidden from the initiator
        required: true
        type: string
      responses:
        '201':
          description: successful operation
...
definitions:
...
  AccountVM:
    type: object
    properties:
      accountId:
        type: string
      displayName:
        type: string
      ethnicity:
        type: string
      dateOfBirth:
        type: string
      gender:
        type: string
      distance:
        type: integer
      pictureUrl:
        type: array
        items:
          type: string
      pictureThumbnailUrl:
        type: array
        items:
          type: string
      signId:
        type: array
        items:
          $ref: '#/definitions/Sign'
  AccountUpdate:
    type: object
    properties:
      signId:
        type: array
        items:
            type: integer
      description:
        type: string
      pictureUrl:
        type: array
        items:
          type: string
...
  Sign:
    type: object
    properties:
      signId:
        type: integer
      signName:
        type: string
      signIconUrl:
        type: string
      order:
        type: integer
...